+++ src/models.rs
@@ -14,7 +14,11 @@
 
 use crate::util::*;
 use chrono::prelude::*;
-use monero::{cryptonote::hash::Hash as CryptoNoteHash, util::address::PaymentId, Address};
+use monero::{
+    cryptonote::{hash::Hash as CryptoNoteHash, subaddress},
+    util::address::PaymentId,
+    Address,
+};
 use serde::{Deserialize, Deserializer, Serialize};
 use std::{collections::HashMap, num::NonZeroU64};
 
@@ -155,7 +159,7 @@ pub struct JsonTransaction {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct SubaddressBalanceData {
     pub address: Address,
-    pub address_index: u64,
+    pub address_index: u32,
     pub balance: u64,
     pub label: String,
     pub num_unspent_outputs: u64,
@@ -201,15 +205,9 @@ pub struct TransferData {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct SubaddressData {
     pub address: Address,
-    pub address_index: u64,
+    pub address_index: u32,
     pub label: String,
     pub used: bool,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct SubaddressIndex {
-    pub major: u64,
-    pub minor: u64,
 }
 
 /// Return type of wallet `get_payments`.
@@ -220,7 +218,7 @@ pub struct Payment {
     pub amount: u64,
     pub block_height: u64,
     pub unlock_time: u64,
-    pub subaddr_index: SubaddressIndex,
+    pub subaddr_index: subaddress::Index,
     pub address: Address,
 }
 
@@ -263,7 +261,7 @@ pub struct IncomingTransfer {
     pub global_index: u64,
     pub key_image: Option<String>,
     pub spent: bool,
-    pub subaddr_index: SubaddressIndex,
+    pub subaddr_index: subaddress::Index,
     pub tx_hash: HashString<CryptoNoteHash>,
     pub tx_size: Option<u64>,
 }
@@ -272,8 +270,8 @@ pub struct IncomingTransfer {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct SweepAllArgs {
     pub address: Address,
-    pub account_index: u64,
-    pub subaddr_indices: Option<Vec<u64>>,
+    pub account_index: u32,
+    pub subaddr_indices: Option<Vec<u32>>,
     pub priority: TransferPriority,
     pub mixin: u64,
     pub ring_size: u64,
@@ -301,8 +299,8 @@ pub struct SweepAllData {
 /// Argument type of wallet `transfer`.
 #[derive(Clone, Debug, Default)]
 pub struct TransferOptions {
-    pub account_index: Option<u64>,
-    pub subaddr_indices: Option<Vec<u64>>,
+    pub account_index: Option<u32>,
+    pub subaddr_indices: Option<Vec<u32>>,
     pub mixin: Option<u64>,
     pub ring_size: Option<u64>,
     pub unlock_time: Option<u64>,
@@ -325,7 +323,7 @@ pub struct GenerateFromKeysArgs {
 /// Return sub-type of wallet `get_accounts`.
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct GotAccount {
-    pub account_index: u64,
+    pub account_index: u32,
     pub balance: u64,
     pub base_address: monero::Address,
     pub label: Option<String>,
@@ -396,9 +394,9 @@ impl From<GetTransfersCategory> for &'static str {
 pub struct GetTransfersSelector {
     pub category_selector: HashMap<GetTransfersCategory, bool>,
     /// Index of the account to query for transfers. (defaults to 0)
-    pub account_index: Option<u64>,
+    pub account_index: Option<u32>,
     /// List of subaddress indices to query for transfers. (Defaults to empty - all indices)
-    pub subaddr_indices: Option<Vec<u64>>,
+    pub subaddr_indices: Option<Vec<u32>>,
     /// Filter transfers by block height.
     pub block_height_filter: Option<BlockHeightFilter>,
 }
@@ -456,7 +454,7 @@ pub struct GotTransfer {
     /// Payment ID for this transfer.
     pub payment_id: HashString<PaymentId>,
     /// JSON object containing the major & minor subaddress index.
-    pub subaddr_index: SubaddressIndex,
+    pub subaddr_index: subaddress::Index,
     /// Estimation of the confirmations needed for the transaction to be included in a block.
     pub suggested_confirmations_threshold: u64,
     /// POSIX timestamp for when this transfer was first confirmed in a block (or timestamp submission if not mined yet).